requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
joblib==1.4.2
nltk==3.8.1
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
pytesseract==0.3.10
regex==2024.5.15
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
